+++ src/python/dxpy/toolkit_version.py
@@ -1,1 +1,1 @@
-version = '0.399.0+g5bfa722be'
+version = '0.399.0+g61cfd6fb0'